add playlistSongs

+++ src/api/playlists/handler.js
@@ -15,6 +15,10 @@ class PlaylistsHandler {
     this.postPlaylistHandler = this.postPlaylistHandler.bind(this);
     this.getPlaylistsHandler = this.getPlaylistsHandler.bind(this);
     this.deletePlaylistByIdHandler = this.deletePlaylistByIdHandler.bind(this);
+    this.addPlaylistSongHandler = this.addPlaylistSongHandler.bind(this);
+    this.getPlaylistSongHandler = this.getPlaylistSongHandler.bind(this);
+    this.deletePlaylistSongHandler =
+     this.deletePlaylistSongHandler.bind(this);
   }
 
   /**
@@ -29,9 +33,7 @@ class PlaylistsHandler {
       const {id: credentialId} = request.auth.credentials;
 
       const playlistId = await this._service.addPlaylist({name,
-        username: credentialId});
-
-      console.log(playlistId);
+        owner: credentialId});
 
       const response = h.response({
         status: 'success',
@@ -58,7 +60,6 @@ class PlaylistsHandler {
         message: 'Maaf, terjadi kegagalan pada server kami.',
       });
       response.code(500);
-      console.error(error);
       return response;
     }
   }
@@ -94,7 +95,6 @@ class PlaylistsHandler {
         message: 'Maaf, terjadi kegagalan pada server kami.',
       });
       response.code(500);
-      console.error(error);
       return response;
     }
   }
@@ -108,6 +108,7 @@ class PlaylistsHandler {
     try {
       const {id: credentialId} = request.auth.credentials;
       const id = request.params['playlistId'];
+      await this._service.verifyPlaylistOwner(id, credentialId);
       await this._service.deletePlaylistById(id, credentialId);
 
       return {
@@ -130,7 +131,123 @@ class PlaylistsHandler {
         message: 'Maaf, terjadi kegagalan pada server kami.',
       });
       response.code(500);
-      console.error(error);
+      return response;
+    }
+  }
+
+
+  /**
+   * @param {string} request
+   * @param {string} h
+   * @return {string} response
+   */
+  async addPlaylistSongHandler(request, h) {
+    try {
+      const {playlistId} = request.params;
+      const {songId} = request.payload;
+      const {id: credentialId} = request.auth.credentials;
+      this._validator.validatePlaylistSongPayload(request.payload);
+
+      await this._service.verifyPlaylistOwner(playlistId, credentialId);
+      await this._service.addPlaylistsong({playlistId, songId});
+
+      const response = h.response({
+        status: 'success',
+        message: 'Lagu berhasil ditambahkan ke playlist',
+      });
+      response.code(201);
+      return response;
+    } catch (error) {
+      if (error instanceof ClientError) {
+        const response = h.response({
+          status: 'fail',
+          message: error.message,
+        });
+        response.code(error.statusCode);
+        return response;
+      }
+
+      // Server ERROR!
+      const response = h.response({
+        status: 'error',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
+      });
+      response.code(500);
+      return response;
+    }
+  }
+
+  /**
+   * @param {string} request
+   * @param {string} h
+   * @return {string} response
+   */
+  async getPlaylistSongHandler(request, h) {
+    try {
+      const {playlistId} = request.params;
+      const {id: credentialId} = request.auth.credentials;
+
+      await this._service.verifyPlaylistOwner(playlistId, credentialId);
+      const songs = await this._service.getPlaylistSongs(playlistId);
+      return {
+        status: 'success',
+        data: {
+          songs,
+        },
+      };
+    } catch (error) {
+      if (error instanceof ClientError) {
+        const response = h.response({
+          status: 'fail',
+          message: error.message,
+        });
+        response.code(error.statusCode);
+        return response;
+      }
+
+      // Server ERROR!
+      const response = h.response({
+        status: 'error',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
+      });
+      response.code(500);
+      return response;
+    }
+  }
+
+  /**
+   * @param {string} request
+   * @param {string} h
+   * @return {string} response
+   */
+  async deletePlaylistSongHandler(request, h) {
+    try {
+      const {playlistId} = request.params;
+      const {songId} = request.payload;
+      const {id: credentialId} = request.auth.credentials;
+
+      await this._service.verifyPlaylistOwner(playlistId, credentialId);
+      await this._service.deletePlaylistSongs(playlistId, songId);
+      return {
+        status: 'success',
+        message: 'Lagu berhasil dihapus dari playlist',
+      };
+    } catch (error) {
+      if (error instanceof ClientError) {
+        const response = h.response({
+          status: 'fail',
+          message: error.message,
+        });
+        response.code(error.statusCode);
+        return response;
+      }
+
+      // Server ERROR!
+      const response = h.response({
+        status: 'error',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
+      });
+      response.code(500);
       return response;
     }
   }
